Ensure tbot compiles on windows (#43877)
* Ensure tbot compiles on windows

* Restore unix tbot support

* fixup! Restore unix tbot support

* address feedback

* Fix tbot build on windows, use CGO for this OS

* Move unix-specific syscalls and file descriptors out of os-agnostic files

* fixup! Move unix-specific syscalls and file descriptors out of os-agnostic files

* lint
